services/llm-engine/pyproject.toml
[project]
name = "llm-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "google-auth>=2.47.0",
    "google-cloud-aiplatform>=1.133.0",
    "google-genai>=1.57.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
